+++ Solutions/16priyesh/2_16priyesh.cpp
