+++ Solutions/AK-007/2_AK-007.cpp
