+++ Solutions/nikki1016/2_nikki1016.cpp
