+++ Solutions/poor-kid/2_poor-kid.cpp
