+++ Solutions/surasnayak/2_surasnayak.cpp
